[🚚rename]: 페이지 컴포넌트의 네이밍을 Page로 명명
