+++ src/pages/MainPage.tsx
@@ -1,30 +1,28 @@
 import { css } from '@emotion/react';
 
-import CodeArchiveList from '@/components/CodeArchive/CodeArchiveList';
+import KanbanBoard from '@/components/KanbanBoard/KanbanBoard';
 import { theme, ThemeType } from '@/styles/theme';
 
-export default function CodeArchive() {
+export default function MainPage() {
   return (
-    <div css={codeArchiveStyle(theme)}>
-      <h1 className="code-archive-title">CODE ARCHIVE</h1>
-      <CodeArchiveList />
-    </div>
+    <main css={mainPageStyle(theme)}>
+      <h1 className="kanban-title">ISSUE BOARD</h1>
+      <KanbanBoard />
+    </main>
   );
 }
 
-const codeArchiveStyle = (theme: ThemeType) => css`
+const mainPageStyle = (theme: ThemeType) => css`
   display: flex;
   flex-direction: column;
   align-items: center;
   overflow: hidden;
-  color: ${theme.colors.text};
 
-  .code-archive-title {
+  .kanban-title {
     font-size: 1.6rem;
     margin-bottom: 10px;
     font-weight: 700;
     color: ${theme.colors.done};
-
     :hover {
       color: ${theme.colors.primary};
     }
