+++ src/routes/index.tsx
@@ -1,20 +1,20 @@
 import { lazy } from 'react';
 
 import Layout from '../pages/_layout';
-import NotFound from '../pages/NotFound';
+import NotFoundPage from '../pages/NotFoundPage';
 import { path } from './path';
 
-const CodeArchive = lazy(() => import('../pages/CodeArchive'));
-const Main = lazy(() => import('../pages/Main'));
+const CodeArchivePage = lazy(() => import('../pages/CodeArchivePage'));
+const MainPage = lazy(() => import('../pages/MainPage'));
 
 export const routes = [
   {
     path: path.root,
     element: <Layout />,
     children: [
-      { index: true, element: <Main /> },
-      { path: path.codeArchive, element: <CodeArchive /> },
-      { path: '*', element: <NotFound /> },
+      { index: true, element: <MainPage /> },
+      { path: path.codeArchive, element: <CodeArchivePage /> },
+      { path: '*', element: <NotFoundPage /> },
     ],
   },
 ];
